Prevented parent navbar from being rerendered
